feat(switchMap): add higher-order lettable version of switchMap
