+++ src/operators/index.ts
@@ -2,3 +2,4 @@ export { concatMap } from './concatMap';
 export { filter } from './filter';
 export { map } from './map';
 export { mergeMap } from './mergeMap';
+export { switchMap } from './switchMap';
